u7: fixed GCC build error

+++ u7/Forms/ChangesetDetailsDialog.cpp
@@ -79,7 +79,7 @@ ChangesetDetailsDialog::ChangesetDetailsDialog(const DbChangesetDetails& changes
 		itemTextList << co.parent();
 
 		QTreeWidgetItem* item = new QTreeWidgetItem(itemTextList);
-		item->setData(0, Qt::UserRole, i);
+		item->setData(0, Qt::UserRole, static_cast<int>(i));
 
 		items.push_back(item);
 	}
